[pgb] abstract pgb api requests
...out of redux actions, into util lib
