+++ src/actions/pgbSessionActions.js
@@ -1,7 +1,4 @@
-/*global cordova*/
-/*global PhonegapBuildOauth*/
-
-import fetch from 'isomorphic-fetch';
+import * as pgb from 'utils/pgb';
 
 export function pgbLoginRequested() {
   return {
@@ -41,105 +38,15 @@ export function pgbAppsReceived(apps) {
   };
 }
 
-function simulateFetchApps(dispatch) {
-  setTimeout(() => {
-    fetch('assets/fakeapps.json')
-    .then(response =>
-      response.json().then((json) => {
-        console.log(`${json.apps.length} apps found`);
-        dispatch(pgbAppsReceived(json.apps));
-      })
-    );
-  }, 2000);
-}
-
-// stub for in browser dev
-function simulateLogin(dispatch) {
-  setTimeout(() => {
-    dispatch(pgbLoginReceived('1234'));
-    //dispatch(pgbAppsRequested());
-    //simulateFetchApps(dispatch);
-  }, 2000);
-}
-
-function getClientID() {
-  return new Promise((resolve, reject) => {
-    PhonegapBuildOauth.getClientID((id) => {
-      resolve(id);
-    });
-  });
-}
-
-function getQueryString(url) {
-  const a = url.slice((url.indexOf('?') + 1)).split('&');
-  if (a === '') return {};
-  const b = {};
-  a.forEach((i) => {
-    const p = i.split('=', 2);
-    if (p.length === 1) {
-      b[p[0]] = '';
-    } else {
-      b[p[0]] = decodeURIComponent(p[1].replace(/\+/g, ' '));
-    }
-  });
-  return b;
-}
-
-function showAuthWindow(clientID) {
-  const authWindow = cordova.InAppBrowser.open(
-    `https://build.phonegap.com/authorize?client_id=${clientID}`,
-    '_blank',
-    'clearcache=yes,location=no'
-  );
-
-  return new Promise((resolve, reject) => {
-    authWindow.addEventListener('loadstart', (e) => {
-      const url = e.url;
-      if (url.match(/^(https?:\/\/)phonegap\.com\/?\?(code|error)=[a-zA-Z0-9_]*$/)) {
-        console.log('Callback url found.');
-        const qs = getQueryString(url);
-        if (qs.code || qs.error) {
-          authWindow.close();
-          resolve(qs.code, qs.error);
-        }
-      }
-    });
-  });
-}
-
-function handleAuth(authCode, err) {
-  return new Promise((resolve, reject) => {
-    if (authCode) {
-      PhonegapBuildOauth.authorizeByCode(authCode, (a) => {
-        resolve(a.access_token);
-      }, (a) => {
-        console.log(`Auth failure: ${a.message}`);
-        alert('Login failed', 'Error');
-        reject();
-      });
-    } else if (err) {
-      console.log(`Auth failure: ${err}`);
-      alert('Login failed', 'Error');
-      reject();
-    }
-  });
-}
-
 export function fetchApps(accessToken) {
   return (dispatch) => {
-    // stub for in browser dev
-    if (typeof cordova === 'undefined' || cordova.platformId === 'browser') {
-      return simulateFetchApps(dispatch);
-    }
+    dispatch(pgbAppsRequested());
 
-    return (fetch(`https://build.phonegap.com/api/v1/apps?access_token=${accessToken}`)
-    .then(response =>
-      response.json().then((json) => {
-        console.log(`${json.apps.length} apps found`);
-        dispatch(pgbAppsReceived(json.apps));
-        return json.apps;
-      })
-    ));
+    return pgb.fetchApps(accessToken)
+    .then((apps) => {
+      console.log(`${apps.length} apps found`);
+      dispatch(pgbAppsReceived(apps));
+    });
   };
 }
 
@@ -147,22 +54,9 @@ export function login() {
   return (dispatch) => {
     dispatch(pgbLoginRequested());
 
-    // stub for in browser dev
-    if (typeof cordova === 'undefined' || cordova.platformId === 'browser') {
-      return simulateLogin(dispatch);
-    }
-
-    return getClientID()
-    .then(showAuthWindow)
-    .then(handleAuth)
+    return pgb.login()
     .then((accessToken) => {
       dispatch(pgbLoginReceived(accessToken));
-      dispatch(pgbAppsRequested());
-      return accessToken;
-    //})
-    //.then(fetchApps)
-    //.then((apps) => {
-      //dispatch(pgbAppsReceived(apps));
     });
   };
 }
